pull MenuItem tests to separate file
